Remove Button interface
The button interface is only used by powerdevil which is normal given
that the functionality it provides is not needed by anybody else.
Nobody needs to know when the power acpi button is press rather than
powerdevil.

Since cross-platform is still a good value we might replace this by
something simpler in Solid::PowerManagement, something that at least
provides the status of the Lid.