Remove Keyboard interface
This interface did not exists in KDE4 and in KF5 was added because
Plasma was going to use it but at the moment it is not using it.

So let's play it safe and remove the interface for the time being, if
later on we decide to add it back that is something we can always do.