Remove Video interface
The only user of it is Kopete and it is using it on code that is
linux-only so it can be ported to use UDev directly.

Also this interface has an additional problem which is that it is
almost useless without a multimedia library that will use the devices
returned by Solid, for example without GStreamer. So since you are
already using a multimedia framework why don't use that framework to
discover video devices?

For a quick port Applications can use QtMultimedia.