Remove AcAdapter interface
This interface is used only by 2 app which check if the AcAdapter is
plugged.

This is a valid use case but having AcAdapter in the Solid devices
does not make sense it you won't have more than 1.

So we will add API for this in Solid::PowerManagement which avoids for
app developers all the hussle of having to query solid, iterate
trhough the returned device list etc.