Move all public api to a frontend folder
It has been brought to my attention that all our source files in
the root of the solid folder are confusing. They where even more
in the kde4 times since they mixed Device inherited classes and some
classes (networking and powermanagement) that had nothing to do with
them.

So, as from this commit we have:
backends/ #Folder where backend code is
ifaces/   #All our interfaces, used on backends
frontend/ #All the classes exported and intended to be used by the user
*	  #Files that are used across folders and other minor stuff