First commit of the Power async api
Build upon Solid::Job, this api will offer a set of jobs that will
provide information regarding the power system, some examples:
-AcPluggedJob
-SleepJob
-SleepStatesJob
...

The backends at the moment will be in source and hardcoded since we
don't have time to write a runtime system for it before release. The
classes though are structured in a way that we will allow backends
to be outside our source tree in the future.

To enable this feature compile with:
-DWITH_NEW_POWER_ASYNC_API=On