Use Q_GLOBAL_STATIC and allow to instance the Power object
There are applications whos live spawn requires them to always be
aware of the power system. A good example of it is Baloo that since
the moment it is executed it needs to be aware of whether the AC is
on or not.

On the other hand, applications like Muon (software installer) they
only need to do power related stuff when they are actually modifying
software (installing, uninstalling, updating).

So, for Baloo we offer a singlethon with QGlobalStatic which is
pretty much memory free.

For Muon we offer our constructor so they can manage the memory
as they please.