Add a proper Inhibition object, with start/stop methods
By popular demand the Inhibition object has now some logic into it.

The idea is that InhibitionJob will return an Inhibition object that
is already started (so the inhibition is in place), and you can keep
that object to start/stop the inhibition depending on the app state.

For example:
The moment a video starts playing in our VideoPlayer we Call
InhibitionJob, hence creating an Inhibition object. If the video
is paused, then we stop the inhibition. When the video is playing
again we start the inhibition.