Rename Power::Inhibitions to Power::States
Instead of having different list of Inhibitions and States we will
instead have a list of supported States by the device that the developer
will be able to set the device to, or to prevent it to happen
(inhibition).

As rightn ow we only support the ones supported in Linux/Freedesktop,
but in the future we will add things only found (at the moment) in
android.