First import of the freedesktop backend
As difference with the past, now we are introducing a freedesktop
backend rather than a "upower/udev/systemd/ separate backends. This
allows us to be more flexible and to make sure we support freedesktop
correctly.

This first import comes with component test that checks how we are
handling dbus and what dbus calls we are doing.