Do not depend on backend job to return the value
It might happen that the backend job is gone (since it will auto delete
after calling emitResult) but the frontend job is still alive. If that
happens and somebody calls isPLugged() we will be asking info from a
deleted pointer.

This patch adds a copy of the result (a bool) plus a isBackendFinished
status to know when the backend job has finished.