Adjust Power::States to reality since we are about to implement them
Until now the states were kinda inherited from the old solid code and we
stpend no time thinking wether they make any sense or not. Now that we
are about to implement them we need to make more sense.

Current states:
Sleep: Sets the device to the preferred sleep method (s2ram, s2disk..)
	It also allows to inhibit any sleep (s2ram, s2disk, hybrid...)
Screen: A set of jobs exits to modify the state of the screen
	(brightnessChange, PowerOffScreen, etc...)
	It allow to inhibit any actions on the screen (brightness,
        dimming, powering it off...)
Reboot: Can be used to reboot the device
	It also can be used to inhibit reboot
PowerOff: Can be used to poweroff the device
	It also can be used to inhibit poweroff