Initial implementation of a KDE Connect backend for Solid
This allows for a tighter integration of KDE Connect with our workspace by providing
battery information directly in the battery monitor, and also enable us to list the
phone in the device notifier like a regular USB thumb drive. The advantage of the latter
is that Dolphin (with a minor change to the predicate query) can automatically update
the Places panel without KDE Connect manually messing with KFilePlaces.

The is based on the initial implementation by Albert Vaca in the kdelibs-kdeconnect
repository. Battery information and device notifier already works, however for some reason
device notifier thinks it's a "Battery". Also it's still rather crashy and synchronous.