Solid::Device::listFromQuery: simplify by avoiding QSet::fromList.
Summary:
1) it makes the order undefined (which breaks kio's kfileplacesmodeltest
on some machines, despite all the measures we took to disable QHash/QSet randomization)
2) set union is a lot of work just to avoid duplicates. Replaced with a single
set of seen elements.

Test Plan:
Now kfileplacesmodeltest fails on my two machines, instead of passing
on one and failing on the other (with the exact same versions of Qt5 and KF5!) :-)
This means the order is stable now, I'll adapt kfileplacesmodeltest to it.

Reviewers: afiestas, ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D1572