Find CPUs by subsystem rather than driver
The driver "processor" is part of the ACPI module, which isn't used by
all systems and architectures.

Detecting devices by CPU subsystem seems to work, and whilst
theoretically could pick up a non CPU device in the same subsystem they
would be filtered out by the the udevmanager which has an extra check
that CPUs are valid and not just empty sockets.

REVIEW: 128083