Fix check that CPU is valid
We have a check that an devices/system/cpu/cpuN entry contains a
processor and not just an empty socket, however it relies on somewhat
outdated /sys files.

cpuN/cpufreq isn't guaranteed to exist as for some systems (my AMD
processor at least).

The docs in the linux kernel imply topology/core_id should always exist,
and should still work as a valid check that we have a populated CPU
socket.

REVIEW: 128085