Work round DBus property fetching bug
Summary:
property() works in a slightly different way to just calling Get().
It allocates the variant of the relevant type in advance in QObject
code, and then calls
the DBus code to populate it.

This fails for QByteArrayList, before it reaches DBus code.
I don't know why, but we need a wokraround in Solid anyway.

From a DBus traffic perspective this code is identical, we just avoid
going through Qt properties.

This shouldn't introduce any new bugs, as we already use GetAll
directly, if anything this brings it more
in line.

BUG: 345871

Test Plan:
Commented out the checkCache so that it always loads data.
Instead of failing, it now works.

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D4305