Add include that is needed in msys2
Summary: In the latest msys2, they updated some windows headers to require this extra include. It's not specific to this project, [openFrameworks had to do it as well](https://github.com/openframeworks/openFrameworks/pull/5184).

Test Plan: None, it just fixes a build. If I could get some guidance I would be willing to try to get a CI build for it.

Reviewers: #frameworks, kfunk

Reviewed By: kfunk

Tags: #frameworks, #windows

Differential Revision: https://phabricator.kde.org/D4778