Don't show an emblem for mounted disks, only unmounted disks
Summary:
The HIG now states that should only be used for non-default, uncommon, unusual, or unexpected  statuses: https://hig.kde.org/components/assistance/emblem.html

Accordingly, we should only show an emblem when a disk is unmounted, not when it's mounted. Being mounted is the expected status for a disk; being unmounted-but-visible-anyway is what's unusual and deserves an emblem.

As an added bonus, this hugely improves the look of the {nav Places} panel by default.

Closes T9740

Test Plan:
Dolphin {nav Places} panel now:
{F6302248, size=full}

Devices Notifier:
{F6303952, size=full}

Reviewers: #vdg, abetts

Reviewed By: #vdg, abetts

Subscribers: broulik, kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T9740

Differential Revision: https://phabricator.kde.org/D15935