Fix monitoring optical drives
We already listen to all change events anyway so rather than explicitly connecting
to a certain path like the old one, we just process everything and try to figure out
if it was meant for an optical drive whose disc became available.

For the lack of an optical drive, I tested it in VirtualBox where attaching and
removing CD images worked fine now.