Use mount point returned from DBus instead of using property value
Summary:
There is a slight time frame (~2ms) where mount has returned but the MountPoints property is not updated. When the mounpoint is queried during that time it will be empty and opening a plugged in USB device in Dolphin will fail.
To avoid this use the mountpoint information returned by the mount() call

Reviewers: broulik

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20433