Port solid from Qt5::Widgets to Qt5::Gui
Summary:
The only usage of widgets was to get the active window. Incorrectly this
was done through QWidget which would fail for any application using just
QtGui (e.g. QtQuick).

As QGuiApplication also provides access to the focus window we can use
QtGui API instead. As a note: the whole thing is X11 specific as a
window ID is passed to another process, which cannot work on Wayland.
It's possible to get it work on Wayland, but that would require:
 * changes to the dbus protocol
 * add a dependency to KWayland
 * or extend QtWayland to generate the required code

Test Plan: Compiles

Reviewers: #frameworks, dhaumann, apol, broulik

Reviewed By: dhaumann, apol, broulik

Subscribers: davidedmundson, dhaumann, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13541