[Fstab] Do not emit deviceAdded twice on fstab/mtab changes
Summary:
The deviceAdded signal was emitted twice, once with the old device list
in m_deviceList, once with the new one.

The first one is actually wrong and was fixed in kdelibs4 by moving the
deviceAdded signal after the m_deviceList update. In KF5, the second
signal was added but the first one never removed.

Move the m_deviceList update to the top (though for deviceRemoved it
does not matter much), and remove the second signal.

Depends on D28779

Test Plan:
1. solid-hardware listen
2. add an entry to the fstab
-> deviceAdded is only emitted once

Reviewers: #frameworks, afiestas, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28782