Add a QString Solid::Device::displayName, used in Fstab Device for network mounts
Summary:
This adds a label field, so that devices can provide a label and a more detailled description.
This label should be used as the device label and description used for subText or tooltips for instance.
The description is used as label by default.

CCBUG: 415281

For use in D26113 and D26114

Reviewers: #frameworks, bruns, sitter

Subscribers: ervin, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28590