Fix imobiledevice backend: Check API version for DEVICE_PAIRED
Add an API-version definition to consumers of the imobiledevice
imported target. The definition is not very fine-grained, but
right now it's enough to distinguish "1.3.0" from "not-1.3.0"
releases. A better approach would unpack the version string
returned from pkg-config into suitable hex values.

The value of IMOBILEDEVICE_API is a (hex) constant consisting
of (major<<16)|(minor<<8)|patch, in order to be compatible
with QT_VERSION_CHECK, so 1.3.0 becomes 0x10300.

In the imobiledevice backend, restore the code that does
nothing with DEVICE_PAIRED, but now protect it with versioning
`#ifdefs` and restore the unhandled-event warning.