autotests: test the predicate-parser
Throw in ten strings for parsing to demonstrate conjunctions
and conjunction nesting; also parse and take-apart a longer
predicate.

Show how predicates apply to devices and that manual filtering
and the list-functions that take predicates do the same thing.
(This test requires there to **be** removable media, so it should
be disabled in CI, among other places)

Demonstrate some of the interface-manipulations (this is based
on internal knowledge of the matches() function), and that nothing
is a volume and a drive at once.

It's ok (e.g. on the CI or on a workstation with no USB drives
plugged in just right now) to have no removable storage attached.
It doesn't affect the rest of the test, since there's only
empty lists (of removable storage) being compared.