Unbreak Device::storageAccessFromPath on Windows
The paths use the Qt/Linux format here with forward slashes. But
the code was adding the native path separator (i.e. backward slash)
on windows, which breaks the storageAccessFromPath logic when
we pass in a path. I.e. when we look for `C:/` we used to turn that
into `C:/\` and then no storage access path would ever match that.