Add Samba's SMB3 filesystem to Solid.
The SMB3 filesystem is now recognized as network filesystem in
_k_isFstabNetworkFileSystem().

When encountering the SMB3 filesystem, FakeNetworkShare::type() is
returning Solid::NetworkShare::Cifs.

Changes to FstabNetworkShare::FstabNetworkShare() are not necessary,
because shares mounted with the SMB3 filesystem are identically formatted
as the ones with the CIFS filesystem. Also, as long as this is not
desired, one does not need to distinguish between these two.