Fix IORegistryEntryGetPath failed on Apple Silicon
On newer macs, the paths for registy entries can be longer than the
512 chars available in io_string_t. IORegistryEntryGetPath would fail
for these devices, and IOKitManager would be unable to enumerate them.
Apple has provided a new API that uses CFString to store the paths,
and is no longer bound by the path length issue. Convert uses of the
old API to the new API so that we can access external storage devices
on Apple Silicon.