Remove invalid Refresh DBus call from UPower backend
The method is only for debugging, and only works when the daemon is
running in debug mode.

Its not obvious why or even if the call was actually required, but
at least with UPower 0.99.18 and later the call no longer works.

(cherry picked from commit f24d846213f41d4e9180ed28164a2ae939d8ec3d)