Avoid cache bypass of `UPowerDevice::allProperties`
The `allProperties()` method always does a synchronous DBus call,
without checking the cache. The method is e.g. used by the
GenericDevice interface to retrieve the properties.

Split the `GetAll` properties call from allProperties, so it can be used
by both `allProperties` and `checkCache` to prime an empty cache.

This also improves the behavior in case the cache was partially populated
by the propertiesChanged signal, now the missing properties are fetched
in bulk instead of individually.