upower: Get rid of QDBusInterface usage
It does a blocking introspect call in its constructor to create
its dynamic properties.

We can just fetch the properties manually using the defines rather
than relying on the interface/service stored in the QDBusInterface.

We can also blindly connect to a signal (install a "filter"),
no need to check whether the interface is valid or not.

While at it, use QDBusReply instead of pending + waitForFinished.