udisks2: Add support CanCheck/Check/CanRepair/Repair
Extend the StorageAccess interface by adding new functions CanCheck/
Check/CanRepair/Repair. Implement their support in udisks2.

This will check the media for file system errors in the plasma-workspace
before mounting it and allow the user to prompt for errors to be
corrected.

This functionality is needed for devicenotifier in plasma-workspace:
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/3758