[DeviceManager] Remove exists() check from storageAccessFromPath
The check is fairly pointless, as the function is typically called with
a path which was valid immediately before, and otherwise the check would
be prone to a TOCTOU race anyway.

It also makes unit testing the function fairly impossible, as the
fakehardware in general has a different filesystem layout than the host,
and we don't want to and can't tie the fakehw to the host filesystem.