[Fstab] Remove remnants of Solaris support
The KDE on Solaris effort is dead for a decade now, remove it. It was also
likely buggy, as the mnttab file was never watched.

Also remove the complicated MNTTAB ifdef'fery, the mtab file name on Linux
is well known, and just set to `/etc/mtab` in the fstabwatcher as well.