fstab: Fix memory leak when a network or overlay mount has changed
Unlike the other functions used in this block, mnt_fs_strdup_options
must be freed manually. While this isn't a serious or obvious issue,
it's known to add up quickly over multiple hours depending on which
filesystems the system is using.

This can end up leaking gigabytes of memory over the course of a day
when the mounts change regularly.