Fix crash when destructing FstabWatcher on Linux
If FstabWatcher encounters an error e.g. kf.solid.backends.fstab: Failed to acquire watch file descriptor Too many open files then mnt_unref_monitor is called. The documetation states:
De-increments reference counter, on zero the mn is automatically deallocated.

Set m_mountMonitor to NULL on failure and check for NULL on destruction In order to not make a double call to the function mnt_unref_monitor which will result in a crash.