backends/fstab: override StorageAccess check signals
The fstab backend overrides the repairRequested and repairDone signals
but does not use them. This means that the frontend can connect to these
signals without having to do a canRepair check beforehand. This is not
the case for checkRequested and checkDone, which print a warning if
trying to connect a signal. For consistency reasons, both check and
repair should behave the same way.

This change overrides the signals for checkRequested and checkDone,
removing the warnings when the frontend tries to connect to them.